add convenience method, remove ord-litecoin API file, cast dynamic->int

+++ lib/services/litescribe_api.dart
@@ -44,7 +44,7 @@ class LitescribeAPI {
     // Check if the number of returned inscriptions equals the limit
     final list = response.data['result']['list'];
     final int total = response.data['result']['total'] as int;
-    final int currentSize = list.length;
+    final int currentSize = list.length as int;
 
     if (currentSize == size && currentSize < total) {
       // If the number of returned inscriptions equals the limit and there are more inscriptions available,
